requirements.txt
django
djangorestframework
# conda install -c conda-forge mysqlclient
mysql-connector-python
mysqlclient